+++ coloraide/interpolate/__init__.py
@@ -19,6 +19,7 @@ import functools
 from abc import ABCMeta, abstractmethod
 from .. import util
 from .. import algebra as alg
+from .. spaces import HSVish, HSLish, Cylindrical, RGBish, LChish, Labish
 from ..types import Vector, ColorInput, Plugin
 from typing import Callable, Sequence, Mapping, Any, TYPE_CHECKING
 
@@ -544,6 +545,71 @@ def normalize_hue(
     return color2, offset
 
 
+def carryforward_convert(color: Color, space: str):
+    """Carry forward undefined values during conversion."""
+
+    cs1 = color._space
+    cs2 = color.CS_MAP[space]
+    carry = []
+    if alg.is_nan(color[-1]):
+        carry.append(-1)
+    channels = {'r': False, 'g': False, 'b': False, 'h': False, 'c': False, 'l': False, 'v': False}
+    if isinstance(cs1, RGBish):
+        for i, name in zip(cs1.rgbish_indexes(), ('r', 'g', 'b')):
+            if alg.is_nan(color[i]):
+                channels[name] = True
+    elif isinstance(cs1, HSLish):
+        for i, name in zip(cs1.hslish_indexes(), ('h', 'c', 'l')):
+            if alg.is_nan(color[i]):
+                channels[name] = True
+    elif isinstance(cs1, HSVish):
+        for i, name in zip(cs1.hsvish_indexes(), ('h', 'c', 'v')):
+            if alg.is_nan(color[i]):
+                channels[name] = True
+    elif isinstance(cs1, LChish):
+        for i, name in zip(cs1.lchish_indexes(), ('l', 'c', 'h')):
+            if alg.is_nan(color[i]):
+                channels[name] = True
+    elif isinstance(cs1, Labish):
+        if alg.is_nan(color[cs1.labish_indexes()[0]]):
+            channels['l'] = True
+    elif isinstance(cs1, Cylindrical):
+        if alg.is_nan(color[cs1.hue_index()]):
+            channels['h'] = True
+
+    if isinstance(cs2, RGBish):
+        indexes = cs2.rgbish_indexes()
+        for e, name in enumerate(('r', 'g', 'b')):
+            if channels[name]:
+                carry.append(indexes[e])
+    elif isinstance(cs2, HSLish):
+        indexes = cs2.hslish_indexes()
+        for e, name in enumerate(('h', 'c', 'l')):
+            if channels[name]:
+                carry.append(indexes[e])
+    elif isinstance(cs2, HSVish):
+        indexes = cs2.hsvish_indexes()
+        for e, name in enumerate(('h', 'c', 'v')):
+            if channels[name]:
+                carry.append(indexes[e])
+    elif isinstance(cs2, LChish):
+        indexes = cs2.lchish_indexes()
+        for e, name in enumerate(('l', 'c', 'h')):
+            if channels[name]:
+                carry.append(indexes[e])
+    elif isinstance(cs2, Labish):
+        indexes = cs2.labish_indexes()
+        if channels['l']:
+            carry.append(indexes[0])
+    elif isinstance(cs2, Cylindrical):
+        if channels['h']:
+            carry.append(cs2.hue_index())
+
+    color.convert(space, in_place=True)
+    for i in carry:
+        color[i] = alg.NaN
+
+
 def interpolator(
     interpolator: str,
     create: type[Color],
@@ -555,6 +621,7 @@ def interpolator(
     premultiplied: bool,
     extrapolate: bool,
     domain: list[float] | None = None,
+    carryforward: bool = True,
     undef: bool = True,
     **kwargs: Any
 ) -> Interpolator:
@@ -582,7 +649,10 @@ def interpolator(
     if out_space is None:
         out_space = current.space()
 
-    current.convert(space, in_place=True)
+    if carryforward:
+        carryforward_convert(current, space)
+    else:
+        current.convert(space, in_place=True)
     offset = 0.0
     hue_index = current._space.hue_index() if hasattr(current._space, 'hue_index') else -1
     normalize_color(current)
@@ -615,7 +685,10 @@ def interpolator(
             stops[i] = None
 
         # Adjust color to space and ensure it fits
-        color = color.convert(space)
+        if carryforward:
+            carryforward_convert(color, space)
+        else:
+            color.convert(space, in_place=True)
         normalize_color(color)
         norm = color[:]
         if hue_index >= 0: