+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,10 @@
 
 ## 2.0
 
+- **NEW**: Interpolation will carry forward undefined channels as specified by the CSS specification. In addition,
+  any HSV-ish color spaces will carry forward hue and saturation to HSL-ish and LCh-ish color spaces. HSV will also
+  carry forward V undefined channels to other HSV-ish color spaces.
+
 - **NEW**: Deprecate `is_nan` if favor of `is_undef`. At some future time, `is_nan` will be removed.
 
 - **NEW**: Rework achromatic normalization. There is now a generic way to perform achromatic detection and normalization